Update the UI for Youth/OrgAdmins to show new Opportunity visual states
